+++ CHANGELOG.md
@@ -29,12 +29,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - dird: fix `purge oldest volume` [PR #1628]
 - Fix continuation on colons in plugin baseclass [PR #1637]
+- plugins: fix cancel handling crash [PR #1595]
 
 [PR #1581]: https://github.com/bareos/bareos/pull/1581
 [PR #1587]: https://github.com/bareos/bareos/pull/1587
 [PR #1589]: https://github.com/bareos/bareos/pull/1589
 [PR #1592]: https://github.com/bareos/bareos/pull/1592
 [PR #1593]: https://github.com/bareos/bareos/pull/1593
+[PR #1595]: https://github.com/bareos/bareos/pull/1595
 [PR #1598]: https://github.com/bareos/bareos/pull/1598
 [PR #1605]: https://github.com/bareos/bareos/pull/1605
 [PR #1606]: https://github.com/bareos/bareos/pull/1606